+++ samples/navmodel-samples/src/main/kotlin/com/bumble/appyx/navmodel/cards/operation/CardsOperation.kt
@@ -1,0 +1,6 @@
+package com.bumble.appyx.navmodel.cards.operation
+
+import com.bumble.appyx.core.navigation.Operation
+import com.bumble.appyx.navmodel.cards.Cards
+
+interface CardsOperation<T> : Operation<T, Cards.State>